feat: implement watermark setting
