+++ frontend/src/components/GeneralSetting/index.ts
@@ -1,0 +1,4 @@
+import BrandingSetting from "./BrandingSetting.vue";
+import WatermarkSetting from "./WatermarkSetting.vue";
+
+export { BrandingSetting, WatermarkSetting };
